changed the ip gathering from data bags to chef api

+++ jenkins/scripts/python/bare_metal/teardown/razor_teardown_policy.py
@@ -123,29 +123,32 @@ else:
         chef_name = "%s%s.%s" % (data['hostname_prefix'], data['bind_number'], data['domain'])
         root_pass = getrootpass(data)
         dbag_uuid = get_data_bag_UUID(data)
+        # get the ip from the data bag
         ip = getip_from_data_bag(dbag_uuid)
-        
-        
 
-        #Remove active model
-        
-        #SSH into ip and reboot 
+        # check to see if box has a chef node and client
+        try:
+            chef_api = ChefAPI(results.chef_url, results.chef_client_pem, results.chef_client)
+            chef_node = chef_api.api_request('GET', '/node/%s' % chef_name)
+            chef_client = chef_api.api_request('GET', '/client/%s' % chef_name)
+            # if chef has a node for this box, change the ip to the ip that chef has for it
+            ip = chef_node['ip_address']
+        except Exception, e:
+            print "Razor node doesnt have a chef node / chef client"
+            continue
         
         if results.display_only == 'true':
             print "Active Model ID: %s " % active
             print "Data Bag UUID: %s " % dbag_uuid
-            #print "ROOT_PASS: %s " % root_pass
             print "Public address: %s " % ip
             print "Private address: %s " % private_ip
             print "Chef Name: %s" % chef_name
             print "Searching chef clients..."
-            try:
-                chef_api = ChefAPI(results.chef_url, results.chef_client_pem, results.chef_client)
-                client = chef_api.api_request('GET', '/clients/%s' % chef_name)
+            if chef_client is not None and chef_node is not None:
+                print "Chef Node: \n %s" % json.dumps(node, indent=4)
                 print "Client: \n %s" % json.dumps(client, indent=4)
-            except Exception, e:
-                print "Error printing chef clients: %s " % e
-                continue
+            else:
+                print "Razor node doesnt have a chef client or node"
         else: 
             print "Removing active model..."
             try:
@@ -156,25 +159,25 @@ else:
                 print "Error removing active model: %s " % e
                 continue
 
-            print "Removing chef-node..."
-            try:
-                 with ChefAPI(results.chef_url, results.chef_client_pem, results.chef_client):
-                    node = Node(chef_name)
+            print "Attempting to remove Chef node %s..." % chef_name
+            if chef_node is not None:
+                try:
+                    print "Removing chef-node %s..." % chef_name
                     node.delete()
-            except Exception, e:
-                print "Error removing chef node: %s " % e
-                continue
+                except Exception, e:
+                    print "Error removing chef node: %s " % e
+                    continue 
 
-            print "Searching chef clients..."
-            try:
-                chef_api = ChefAPI(results.chef_url, results.chef_client_pem, results.chef_client)
-                response = chef_api.api_request('DELETE', '/clients/%s' % chef_name)
-                print "Client %s removed with response: %s" % (chef_name, response)
-            except Exception, e:
-                print "Error removing chef node: %s " % e
-                continue
+            print "Attempting to remove chef client %s..." % chef_name
+            if chef_client is not None:
+                try:
+                    response = chef_api.api_request('DELETE', '/clients/%s' % chef_name)
+                    print "Client %s removed with response: %s" % (chef_name, response)
+                except Exception, e:
+                    print "Error removing chef client: %s " % e
+                    continue
             
-            print "Trying restart...."
+            print "Trying to restart server %s with ip %s...." % (chef_name, ip)
             try:
                 subprocess.call("sshpass -p %s ssh -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no -o LogLevel=quiet -l root %s 'reboot 0'" % (root_pass, ip), shell=True)
                 print "Restart success."
